requirements.txt
Django==1.7.1
Minidetector==0.1
MySQL-python==1.2.5
beautifulsoup4==4.3.2
dj-database-url==0.2.2
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-grappelli==2.5.3
django-security==0.1.22b
django-tastypie==0.12.0
django-view-cache-utils==0.1.1
docutils==0.11
gunicorn==18.0
lxml==3.3.4
psycopg2==2.5.1
pystache==0.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
sqlparse==0.1.13
static==1.1.1
static3==0.5.1
wsgiref==0.1.2
